+++ dorado/decode/GPUDecoder.cpp
@@ -16,9 +16,9 @@ namespace dorado {
 at::Tensor GPUDecoder::gpu_part(at::Tensor scores, int num_chunks, DecoderOptions options) {
     c10::cuda::CUDAGuard device_guard(scores.device());
     utils::ScopedProfileRange loop{"gpu_decode", 1};
-    long int N = scores.sizes()[0];
-    long int T = scores.sizes()[1];
-    long int C = scores.sizes()[2];
+    long int N = (long int)(scores.sizes()[0]);
+    long int T = (long int)(scores.sizes()[1]);
+    long int C = (long int)(scores.sizes()[2]);
 
     auto tensor_options_int32 =
             at::TensorOptions().dtype(at::kInt).device(scores.device()).requires_grad(false);
@@ -52,7 +52,7 @@ at::Tensor GPUDecoder::gpu_part(at::Tensor scores, int num_chunks, DecoderOption
                 chunks.data_ptr(), chunk_results.data_ptr(), N, scores.data_ptr(),
                 m_score_clamp_val, C, aux.data_ptr(), path.data_ptr(), moves.data_ptr(), NULL,
                 sequence.data_ptr(), qstring.data_ptr(), options.q_scale, options.q_shift,
-                options.beam_width, options.beam_cut, options.blank_score));
+                int(options.beam_width), options.beam_cut, options.blank_score));
     }
     {
         utils::ScopedProfileRange spr{"beam_search", 2};
@@ -60,7 +60,7 @@ at::Tensor GPUDecoder::gpu_part(at::Tensor scores, int num_chunks, DecoderOption
                 chunks.data_ptr(), chunk_results.data_ptr(), N, scores.data_ptr(),
                 m_score_clamp_val, C, aux.data_ptr(), path.data_ptr(), moves.data_ptr(), NULL,
                 sequence.data_ptr(), qstring.data_ptr(), options.q_scale, options.q_shift,
-                options.beam_width, options.beam_cut, options.blank_score));
+                int(options.beam_width), options.beam_cut, options.blank_score));
     }
     {
         utils::ScopedProfileRange spr{"compute_posts", 2};
@@ -68,7 +68,7 @@ at::Tensor GPUDecoder::gpu_part(at::Tensor scores, int num_chunks, DecoderOption
                 chunks.data_ptr(), chunk_results.data_ptr(), N, scores.data_ptr(),
                 m_score_clamp_val, C, aux.data_ptr(), path.data_ptr(), moves.data_ptr(), NULL,
                 sequence.data_ptr(), qstring.data_ptr(), options.q_scale, options.q_shift,
-                options.beam_width, options.beam_cut, options.blank_score));
+                int(options.beam_width), options.beam_cut, options.blank_score));
     }
     {
         utils::ScopedProfileRange spr{"decode", 2};
@@ -76,7 +76,7 @@ at::Tensor GPUDecoder::gpu_part(at::Tensor scores, int num_chunks, DecoderOption
                 chunks.data_ptr(), chunk_results.data_ptr(), N, scores.data_ptr(),
                 m_score_clamp_val, C, aux.data_ptr(), path.data_ptr(), moves.data_ptr(), NULL,
                 sequence.data_ptr(), qstring.data_ptr(), options.q_scale, options.q_shift,
-                options.beam_width, options.beam_cut, options.blank_score, options.move_pad));
+                int(options.beam_width), options.beam_cut, options.blank_score, options.move_pad));
     }
     return moves_sequence_qstring.reshape({3, N, -1});
 }
@@ -87,8 +87,8 @@ std::vector<DecodedChunk> GPUDecoder::cpu_part(at::Tensor moves_sequence_qstring
     auto moves_cpu = moves_sequence_qstring_cpu[0];
     auto sequence_cpu = moves_sequence_qstring_cpu[1];
     auto qstring_cpu = moves_sequence_qstring_cpu[2];
-    int N = moves_cpu.size(0);
-    int T = moves_cpu.size(1);
+    int N = int(moves_cpu.size(0));
+    int T = int(moves_cpu.size(1));
 
     std::vector<DecodedChunk> called_chunks;
